+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add module for applying default service quotas.
+
 ## [7.2.0] - 2025-11-28
 
 ### Added